Oppnår målet ved å få flytta brev-klassa over til modellen
